start of comp layout
