+++ nems/gui_new/ui/container.ui
@@ -25,7 +25,7 @@
         <string>Browser</string>
        </attribute>
       </widget>
-      <widget class="QWidget" name="tabComp">
+      <widget class="CompTab" name="tabComp">
        <attribute name="title">
         <string>Comp</string>
        </attribute>
@@ -90,6 +90,12 @@
    <class>BrowserTab</class>
    <extends>QWidget</extends>
    <header>browser_tab</header>
+   <container>1</container>
+  </customwidget>
+  <customwidget>
+   <class>CompTab</class>
+   <extends>QWidget</extends>
+   <header>comp_tab</header>
    <container>1</container>
   </customwidget>
  </customwidgets>
