+++ nems/gui_new/ui_promoted.py
@@ -2,6 +2,72 @@ from PyQt5.QtCore import *
 from PyQt5.QtGui import *
 from PyQt5.QtWidgets import *
 
+from nems import db
+
+
+class ListViewModel(QAbstractListModel):
+    def __init__(self,
+                 table_name: str,
+                 column_name: str,
+                 filter=None,
+                 *args,
+                 **kwargs
+                 ):
+        super(ListViewModel, self).__init__(*args, **kwargs)
+
+        self.table_name = table_name
+        self.column_name = column_name
+
+        self.data = None
+
+        self.init_db()
+        self.refresh_data(filters=filter)
+
+    def data(self, index, role):
+        if role == Qt.DisplayRole:
+            text = self.data[index.row()][0]
+            return text
+
+    def rowCount(self, index):
+        return len(self.data)
+
+    def init_db(self):
+        self.engine = db.Engine()
+        table = db.Tables()[self.table_name]
+        self.column = getattr(table, self.column_name)
+        self.session = db.sessionmaker(bind=self.engine)()
+
+    def close_db(self):
+        self.session.close()
+        self.engine.dispose()
+
+    def refresh_data(self, filters: dict = None):
+        """Reads the data from the database, optionally filtering.
+
+        :param filters: Dict of column: filter.
+        :return:
+        """
+        query = (self.session
+                 .query(self.column)
+                 .distinct()
+                 )
+
+        if filters is not None:
+            query = query.filter_by(**filters)
+
+        query = query.order_by(self.column)
+        data = query.all()
+
+        self.data = data
+
+    def get_index(self, value) -> int:
+        """Searches for a value in the data, returns the index"""
+        flat = [i[0] for i in self.data]
+        try:
+            return flat.index(value)
+        except ValueError:
+            return -1
+
 
 class ExtendedComboBox(QComboBox):
     def __init__(self, parent=None):
